// src/types/enum/AJVSchemaEnum.ts
export enum AJVSchemaEnum {
    CompareCertReq = 'CompareCertReq',
    RepairOOSAccountsReq = 'RepairOOSAccountsReq',
    RequestStateForTxReq = 'RequestStateForTxReq',
    RequestStateForTxResp = 'RequestStateForTxResp',
    RequestReceiptForTxReq = 'RequestReceiptForTxReq',
    RequestReceiptForTxResp = 'RequestReceiptForTxResp',
    GetAppliedVoteReq = 'GetAppliedVoteReq',
    GetAppliedVoteResp = 'GetAppliedVoteResp',
    LostReportReq = 'LostReportReq',
    RequestStateForTxPostReq = 'RequestStateForTxPostReq',
    RequestStateForTxPostResp = 'RequestStateForTxPostResp',
    WrappedDataResponse = 'WrappedDataResponse',
    BroadcastStateReq = 'BroadcastStateReq',
    GetAccountDataReq = 'GetAccountDataReq',
    GetAccountDataResp = 'GetAccountDataResp',
    ApoptosisProposalReq = 'ApoptosisProposalReq',
    ApoptosisProposalResp = 'ApoptosisProposalResp',
    GetCachedAppDataReq = 'GetCachedAppDataReq',
    GetCachedAppDataResp = 'GetCachedAppDataResp',
    GetAccountDataByListReq = 'GetAccountDataByListReq',
    GetAccountDataByListResp = 'GetAccountDataByListResp',
    GetTrieAccountHashesReq = 'GetTrieAccountHashesReq',
    GetTrieAccountHashesResp = 'GetTrieAccountHashesResp',
    WrappedDataFromQueueSerializable = 'WrappedDataFromQueueSerializable',
    GetAccountDataWithQueueHintsReq = 'GetAccountDataWithQueueHintsReq',
    GetAccountDataWithQueueHintsResp = 'GetAccountDataWithQueueHintsResp',
    SpreadAppliedVoteHashReq = 'SpreadAppliedVoteHashReq',
    GlobalAccountReportReq = 'GlobalAccountReportReq',
    GlobalAccountReportResp = 'GlobalAccountReportResp',
    SyncTrieHashesRequest = 'SyncTrieHashesRequest',
    SyncTrieHashesReq = 'SyncTrieHashesReq',
    GetAccountQueueCountReq = 'GetAccountQueueCountReq',
    GetAccountQueueCountResp = 'GetAccountQueueCountResp',
    MakeReceiptReq = 'MakeReceiptReq',
    GetTrieHashesReq = 'GetTrieHashesReq',
    GetTrieHashesResp = 'GetTrieHashesResp',
    SignAppDataReq = 'SignAppDataReq',
    SignAppDataResp = 'SignAppDataResp',
    CachedAppData = 'CachedAppData',
    SendCachedAppDataReq = 'SendCachedAppDataReq',
    SpreadTxToGroupSyncingReq = 'SpreadTxToGroupSyncingReq',
    CompareCertResp = 'CompareCertResp',
    GetAccountDataByHashesReq = 'GetAccountDataByHashesReq',
    GetTxTimestampReq = 'GetTxTimestampReq',
    GetTxTimestampResp = 'GetTxTimestampResp',
    GetAccountDataByHashesResp = 'GetAccountDataByHashesResp',
    CycleRecords = 'CycleRecords',
    JoinReq = 'JoinReq'
}